+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ### Added
 * add `show`option to show the code block in terminal
- 
+* show `<lang>: not configured` next to the code block name
+
 ### Changed
  
 ### Fixed